feat(events): remove the standalone outbox relay
